projects/neo-bot/pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "repl_python3_beta_GrizzledExternalForce"
version = "0.1.0"
[tool.poetry.dependencies]
asyncpraw = "7.2.0"
discord-components = "0.5.2.4"
discord-py-slash-command = "1.2.0"
"discord.py" = "^1.7.2"
python = "^3.8"
pyyaml = "5.3.1"
wikipedia = "1.4.0"
[tool.poetry.dev-dependencies]
